ENH: data_api --> data_catalog, data_providers --> data_provider
