--- adlkit/data_api/config.py
@@ -1,4 +1,0 @@
-import os
-
-base_dir = os.path.join(os.path.expanduser('~/'), 'dmz')
-# label_dir = 'labels'
