+++ adlkit/data_catalog/old_abstract_data_catalog.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 import logging as lg
 
-from adlkit.data_api.utils import epoch_time_to_file_name, timestamp_to_epoch
+from adlkit.data_catalog.utils import epoch_time_to_file_name, timestamp_to_epoch
 
 
 class Label(object):
