+++ adlkit/data_catalog/old_base_data_catalog.py
@@ -1,8 +1,8 @@
 import os
 import shelve
 
-from adlkit.data_api.core import DataAPI
-from adlkit.data_api.utils import epoch_time_to_file_name, file_name_to_epoch_time
+from adlkit.data_catalog.old_abstract_data_catalog import DataAPI
+from adlkit.data_catalog.utils import epoch_time_to_file_name, file_name_to_epoch_time
 
 
 class FileDataAPI(DataAPI):
