+++ adlkit/data_provider/bin/gen_rand_data.py
