+++ setup.py
@@ -2,7 +2,7 @@ import setuptools
 
 setuptools.setup(
     name='adlkit',
-    version='0.2',
+    version='0.3',
     author='Amir Farbin, William Hilliard, Ryan Reece',
     author_email='wghilliard@anomalousdl.com',
     url='https://github.com/anomalousdl/adlkit',
